+++ app/controllers/identifiers_controller.rb
@@ -118,7 +118,7 @@ class IdentifiersController < ApplicationController
 
   # GET /identifiers/download
   def download
-    send_data Export::Download.generate_csv(Identifier.where(project_id: sessions_current_project_id)), type: 'text', filename: "identifiers_#{DateTime.now}.csv"
+    send_data Export::Download.generate_csv(Identifier.where(project_id: sessions_current_project_id)), type: 'text', filename: "identifiers_#{DateTime.now}.tsv"
   end
 
   # GET /identifiers/identifier_types