+++ app/controllers/ranged_lot_categories_controller.rb
@@ -72,7 +72,7 @@ class RangedLotCategoriesController < ApplicationController
   def download
     send_data(Export::Download.generate_csv(RangedLotCategory.where(project_id: sessions_current_project_id)),
               type: 'text',
-              filename: "ranged_lot_categories_#{DateTime.now}.csv")
+              filename: "ranged_lot_categories_#{DateTime.now}.tsv")
   end