+++ app/controllers/tasks/projects/data_controller.rb
@@ -7,12 +7,12 @@ class Tasks::Projects::DataController < ApplicationController
   end
 
   def sql_download
-    download = ::Export::ProjectData::Sql.download(sessions_current_project)
+    download = ::Export::ProjectData::Sql.download_async(sessions_current_project)
     redirect_to download_path(download)
   end
 
   def tsv_download
-    download = ::Export::ProjectData::Tsv.download_async(Project.find(sessions_current_project_id))
+    download = ::Export::ProjectData::Tsv.download_async(sessions_current_project)
     redirect_to file_download_path(download)
   end