+++ app/controllers/taxon_determinations_controller.rb
@@ -113,7 +113,7 @@ class TaxonDeterminationsController < ApplicationController
   def download
     send_data Export::Download.generate_csv(TaxonDetermination.where(project_id: sessions_current_project_id)),
       type: 'text',
-      filename: "taxon_determinations_#{DateTime.now}.csv"
+      filename: "taxon_determinations_#{DateTime.now}.tsv"
   end
 
   # POST /taxon_determinations/batch_create