+++ app/controllers/taxon_names_controller.rb
@@ -108,7 +108,7 @@ class TaxonNamesController < ApplicationController
   def download
     send_data Export::Download.generate_csv(
       TaxonName.where(project_id: sessions_current_project_id)
-    ), type: 'text', filename: "taxon_names_#{DateTime.now}.csv"
+    ), type: 'text', filename: "taxon_names_#{DateTime.now}.tsv"
   end
 
   def batch_load