+++ app/models/download/project_dump/tsv.rb
@@ -1,2 +1,2 @@
-class ::Download::ProjectDump::Tsv < ::Download::ProjectDump
+class Download::ProjectDump::Tsv < Download::ProjectDump
 end