+++ lib/export/project_data/tsv.rb
@@ -1,54 +1,46 @@
+# !!
+# !! Only exports ApplicationEnumeration.project_data_classes
+# !! See the .sql version for code that would handle Community data
+# !!
 module ::Export::ProjectData::Tsv
 
-  # Restorable with psql -U :username -d :database -f dump.sql. Requires database to be created without tables (rails db:create)
-  def self.generate_dump(project, file)
-    config = ActiveRecord::Base.connection_config
+  # @return path
+  #   to the zipped data in /tmp
+  def self.export(project)
+    zip_file_path = "/tmp/_#{SecureRandom.hex(8)}_project_tsv.zip"
 
-    args = [
-      ['-h', config[:host]],
-      ['-s', config[:database]],
-      ['-U', config[:username]],
-      ['-p', config[:port]]
-    ].reject! { |(a, v)| v.nil? }.flatten!
-
-    # Retrieve schema
-    schema = Open3.capture3({'PGPASSWORD' => config[:password]}, 'pg_dump', '-w', '-O', *args).first
-
-    # Open gap to insert COPY statements (between tables creation and contraints & indices setup)
-    split_point = schema.index(/\n(--[^\n]*\n)*\s*ALTER\s+TABLE/)
-    schema_head, schema_tail = schema[0..split_point-1], schema[split_point..-1]
-
-    tables = (ActiveRecord::Base.connection.tables - Export::Project::SPECIAL_TABLES - Export::Project::HIERARCHIES.map(&:first)).sort
-
-    file.puts schema_head
-    file.write "\n-- DATA RESTORE\n\n"
-    export_users(file, project)
-
-    tables.each { |t| dump_table(t, file, project.id) }
-    tables.each { |t| setup_pk_sequence(t, file) if get_table_cols(t).include?('"id"') }
-    Export::Project::HIERARCHIES.each { |p| dump_hierarchy_table(p, file, project.id) }
-
-    file.puts schema_tail
+    Zip::File.open(zip_file_path, Zip::File::CREATE) do |zipfile|
+      ApplicationEnumeration.project_data_classes.each do |k|
+        q = k.where(project_id: project.to_param)
+        next if !q.any?
+        copy_table(k, zipfile, q)
+      end
+    end
+    zip_file_path
   end
 
-  # Return path to the data itself
-  def self.export(project)
-    Tempfile.create do |file|
-      buffer = ::Zip::OutputStream.write_buffer(file) do |zipfile|
-        ApplicationEnumeration.project_data_classes.each do |k|
-          q = k.where(project_id: project.to_param)
-          next if !q.any?
-          filename = Zaru::sanitize!("#{k.name}_#{DateTime.now}.tsv")
-          zipfile.put_next_entry(filename)
-          copy_table(k, zipfile, q )
-        end
-      end
-    buffer.flush
-    file.rewind
-    file.path
+  def self.copy_table(klass, zipfile, query)
+    Tempfile.create do |table_file|
+      query = "COPY ( #{query.to_sql} ) TO '#{table_file.path}' WITH (FORMAT CSV, DELIMITER E'\t', HEADER, ENCODING 'UTF8')"
+
+      klass.connection.execute(query)
+
+      table_file.flush
+      table_file.rewind
+      filename = Zaru::sanitize!("#{klass.name}_#{DateTime.now}.tsv")
+
+      # ?! Does not queue the reads until the zipfile loop closes,
+      # so tempfiles may dispappear before they can be read
+      #
+      # zipfile.add(filename, table_file.path )
+
+      zipfile.get_output_stream(filename) { |f| f.write table_file.read }
     end
   end
 
+  # TODO - DRY with generic params passing
+  # @return Download
+  #   a completely built Download, but unsaved
   def self.download(target_project)
     file_path = export(target_project)
 
@@ -57,110 +49,22 @@ module ::Export::ProjectData::Tsv
     d
   end
 
-  def self.copy_table(klass, target_io, query)
-    Tempfile.create do |table_file|
-      query = "COPY ( #{query.to_sql} ) TO '#{table_file.path}' WITH (FORMAT CSV, DELIMITER E'\t', HEADER, ENCODING 'UTF8')"
-
-      klass.connection.execute(query)
-
-      table_file.flush
-      table_file.rewind
-
-      target_io.write table_file.read
-      table_file.close
-    end
-  end
-
+  # @return Download
   def self.download_async(target_project)
-    DownloadProjectTsvJob.perform_later(target_project, stub_download(target_project))
-    download
+    d = stub_download(target_project)
+    DownloadProjectTsvJob.perform_later(target_project, d)
+    d
   end
 
+  # @return Download
+  #   everything except source_file_path
   def self.stub_download(target_project)
-
-    t = Download::ProjectDump::Tsv.new(
+    Download::ProjectDump::Tsv.create!(
        name: "#{target_project.name} export on #{Time.current}.",
        description: 'A zip file containing TSV dump of project-specific data.',
        filename: Zaru::sanitize!("#{target_project.name}_tsv.zip").gsub(' ', '_').downcase,
        expires: 2.days.from_now,
        is_public: false
      )
-     byebug
-     t
   end
-
-=begin
-  def self.setup_pk_sequence(table, io)
-    io.write("SELECT setval('public.#{table}_id_seq', (SELECT COALESCE(MAX(id), 0)+1 FROM public.#{table}));\n\n")
-  end
-
-  def self.get_connection
-    ActiveRecord::Base.connection.raw_connection
-  end
-
-  def self.get_table_cols(table)
-    ActiveRecord::Base.connection.columns(table).map { |c| "\"#{c.name}\"" }
-  end
-
-  def self.dump_table(table, io, project_id)
-    cols = get_table_cols(table)
-    if cols.include?('"project_id"')
-      where_clause = "WHERE project_id IN (#{project_id}, NULL)"
-    elsif table == 'projects'
-      where_clause = "WHERE id = #{project_id}"
-    else
-      where_clause = ''
-    end
-
-    copy_table(table, io, cols, "SELECT #{cols.join(', ')} FROM #{table} #{where_clause}")
-  end
-
-  def self.dump_hierarchy_table(table_pair, io, project_id)
-    cols_model = get_table_cols(table_pair.second)
-
-    return dump_table(table_pair.first, io, project_id) unless cols_model.include?('"project_id"')
-
-    cols_hierarchy = get_table_cols(table_pair.first)
-    select_query = "SELECT #{cols_hierarchy.map { |c| "#{table_pair.first}.#{c}" }.join(', ')} "\
-                   "FROM #{table_pair.second} INNER JOIN "\
-                       "#{table_pair.first} ON #{table_pair.second}.id IN ("\
-                         "#{table_pair.first}.ancestor_id, #{table_pair.first}.descendant_id"\
-                       ') '\
-                   "WHERE project_id = #{project_id}"
-    copy_table(table_pair.first, io, cols_hierarchy, select_query)
-  end
-
-  def self.export_users(io, project)
-    members = project.users.all
-    conn = get_connection
-
-    User.all.each do |user|
-      attributes = {
-        id: user.id,
-        password_digest: "'#{conn.escape_string(User.new(password: 'taxonworks').password_digest)}'",
-        created_at: "'#{user.created_at}'",
-        updated_at: "'#{user.updated_at}'",
-        created_by_id: user.created_by_id || 'NULL',
-        updated_by_id: user.updated_by_id || 'NULL',
-        is_administrator: user.is_administrator || 'NULL',
-        hub_tab_order: "'{#{conn.escape_string(user.hub_tab_order.join(','))}}'"
-      }.merge!(
-        if members.include?(user)
-          {
-            email: "'#{conn.escape_string(user.email)}'",
-            name: "'#{conn.escape_string(user.name)}'"
-          }
-        else
-          {
-            email: "'unknown_#{user.id}@example.com'",
-            name: "'User outside project (#{user.id})'"
-          }
-        end
-      )
-      io.puts("INSERT INTO public.users(#{attributes.keys.join(', ')})\nVALUES (#{attributes.values.join(', ')});")
-    end
-    io.puts
-    setup_pk_sequence('users', io)
-  end
-=end
 end